+++ read_apis/database.go
@@ -1,0 +1,106 @@
+package main
+
+import (
+	"context"
+	"database/sql"
+	"fmt"
+	"io"
+	"log"
+	"net"
+	"os"
+
+	"cloud.google.com/go/cloudsqlconn"
+	"cloud.google.com/go/storage"
+	sql_driver "github.com/go-sql-driver/mysql"
+	"github.com/packit461/packit23/read_apis/models"
+)
+
+func connect_test_db() (*sql.DB, error) {
+	db, err := sql.Open(
+		"mysql",
+		"db_user:oldpassword!!!@tcp(127.0.0.1:3306)/test_db",
+	)
+	if err != nil {
+		log.Fatal(err)
+		return nil, fmt.Errorf("sql.Open: %v", err)
+	}
+	return db, nil
+}
+
+func connect() (*sql.DB, error) {
+	mustGetenv := func(k string) string {
+		v := os.Getenv(k)
+		if v == "" {
+			log.Fatalf("Fatal Error in connect_connector.go: %s environment variable not set.", k)
+		}
+		return v
+	}
+
+	var (
+		db_user                  = mustGetenv("DB_USER")
+		db_pass                  = mustGetenv("DB_PASSWORD")
+		db_name                  = mustGetenv("DB_NAME")
+		instance_connection_name = mustGetenv("INSTANCE_CONNECTION_NAME")
+		usePrivate               = os.Getenv("PRIVATE_IP")
+	)
+
+	d, err := cloudsqlconn.NewDialer(context.Background())
+	if err != nil {
+		return nil, fmt.Errorf("cloudsqlconn.NewDialer: %v", err)
+	}
+	var opts []cloudsqlconn.DialOption
+	if usePrivate != "" {
+		opts = append(opts, cloudsqlconn.WithPrivateIP())
+	}
+
+	sql_driver.RegisterDialContext("cloudsqlconn",
+		func(ctx context.Context, addr string) (net.Conn, error) {
+			return d.Dial(ctx, instance_connection_name, opts...)
+		})
+
+	dbURI := fmt.Sprintf("%s:%s@cloudsqlconn(localhost:3306)/%s?parseTime=true",
+		db_user, db_pass, db_name)
+
+	dbPool, err := sql.Open("mysql", dbURI)
+	if err != nil {
+		return nil, fmt.Errorf("sql.Open: %v", err)
+	}
+	return dbPool, nil
+}
+
+func getBucketObject(bucketName string, objectName string) ([]byte, error) {
+	ctx := context.Background()
+	client, err := storage.NewClient(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("storage.NewClient: %v", err)
+	}
+	defer client.Close()
+	rc, err := client.Bucket(bucketName).Object(objectName).NewReader(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("Object(%q).NewReader: %v", objectName, err)
+	}
+	defer rc.Close()
+	bucketObject, err := io.ReadAll(rc)
+	if err != nil {
+		return nil, fmt.Errorf("ioutil.ReadAll: %v", err)
+	}
+	return bucketObject, nil
+}
+
+// return all versions of package name in db
+func getMetadataFromName(db *sql.DB, name string) ([]models.PackageMetadata, error) {
+	var metadataList []models.PackageMetadata
+	rows, err := db.Query("SELECT id, name, version FROM packages WHERE name = ?;", name)
+	if err != nil {
+		log.Print(err)
+	}
+	defer rows.Close()
+	for rows.Next() {
+		var md models.PackageMetadata
+		if err := rows.Scan(&md.ID, &md.Name, &md.Version); err != nil {
+			return nil, fmt.Errorf("version of package not found. rows.Scan: %v", err)
+		}
+		metadataList = append(metadataList, md)
+	}
+	return metadataList, nil
+}
