--- read_apis/src/read_test.go
@@ -1,16 +1,0 @@
-package main
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestRegister(t *testing.T) {
-	// newUser := model.AuthenticationInput{
-	// 	Username: "yemiwebby",
-	// 	Password: "test",
-	// }
-	// writer := makeRequest("POST", "/auth/register", newUser, false)
-	assert.Equal(t, "test", "test")
-}
